chore: better generator protocol support
